+++ src/ir_LG.hpp
@@ -133,7 +133,7 @@ void IRsend::sendLG2Repeat() {
     mark(LG2_HEADER_MARK);          // + 3000
     space(LG_REPEAT_HEADER_SPACE);  // - 2250
     mark(LG_BIT_MARK);              // + 500
-#if !defined(DISABLE_RECEIVER_RESTART_AFTER_SENDING)
+#if !defined(DISABLE_CODE_FOR_RECEIVER)
     IrReceiver.restartAfterSend();
 #endif
 }
@@ -147,7 +147,7 @@ void sendLG2SpecialRepeat() {
     IrSender.mark(LG2_HEADER_MARK);          // + 3000
     IrSender.space(LG_REPEAT_HEADER_SPACE);  // - 2250
     IrSender.mark(LG_BIT_MARK);              // + 500
-#if !defined(DISABLE_RECEIVER_RESTART_AFTER_SENDING)
+#if !defined(DISABLE_CODE_FOR_RECEIVER)
     IrReceiver.restartAfterSend();
 #endif
 }
@@ -348,7 +348,7 @@ void IRsend::sendLG(unsigned long data, int nbits) {
 // Data + stop bit
     sendPulseDistanceWidthData(LG_BIT_MARK, LG_ONE_SPACE, LG_BIT_MARK, LG_ZERO_SPACE, data, nbits, PROTOCOL_IS_MSB_FIRST,
             SEND_STOP_BIT);
-#if !defined(DISABLE_RECEIVER_RESTART_AFTER_SENDING)
+#if !defined(DISABLE_CODE_FOR_RECEIVER)
     IrReceiver.restartAfterSend();
 #endif
 }